pyproject.toml
[project]
name = "cs-agent-service"
version = "0.1.0"
description = "Alma Coder - Google ADK based agent service"
authors = [
    {name = "Alma Coder Team"}
]
dependencies = [
    "openai",
    "a2a-sdk[http-server]>=0.3.0",
    "httpx>=0.27.0",
    "uvicorn",
    "google-adk",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=src --cov-report=term-missing"
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"

[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "**/__init__.py",
    "src/agent.py",
    "src/mcp_config.py",
    "src/tracing.py",
    "src/__main__.py"
]

[tool.coverage.report]
show_missing = true
fail_under = 50

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
